Revert "Develop UI top bar (#6)"
This reverts commit 4f8de751883e9f7929396025f37f4f8ba0ef8864.
