--- ui/src/components/features/common/index.ts
@@ -1,1 +1,0 @@
-export * from "./AccountMenu";
