+++ ui/src/components/features/dashboard/index.ts
@@ -1,1 +1,2 @@
+export * from "./AccountMenu";
 export * from "./Wall";
