+++ ui/src/components/icons/index.ts
@@ -5,8 +5,5 @@ export {
   Edit as EditIcon,
   Favorite as FavoriteIcon,
   FavoriteBorder as FavoriteBorderIcon,
-  Logout as SignoutIcon,
   MoreVert as MoreVertIcon,
-  Person as PersonIcon,
-  Settings as SettingsIcon,
 } from "@mui/icons-material";
