--- ui/src/components/presentational/Avatar.tsx
@@ -1,8 +1,0 @@
-import { Avatar, AvatarProps } from "@mui/material";
-import { FC } from "react";
-
-const StyledAvatar: FC<AvatarProps> = ({ children, ...props }) => {
-  return <Avatar {...props}>{children}</Avatar>;
-};
-
-export { StyledAvatar as Avatar };
