+++ ui/src/components/presentational/Dialog.tsx
@@ -15,7 +15,6 @@ import { FC } from "react";
 const StyledDialog: FC<DialogProps> = ({ children, ...props }) => {
   return (
     <Dialog
-      disableScrollLock={true}
       sx={{
         "& .MuiDialog-paper": {
           minWidth: "500px",
