+++ ui/src/components/presentational/index.ts
@@ -1,5 +1,4 @@
 export * from "./Alert";
-export * from "./Avatar";
 export * from "./Box";
 export * from "./Button";
 export * from "./Card";
